Reorganised DaoModule
