+++ src/app/dao/server-endpoint/dao-election.ts
@@ -1,12 +1,12 @@
-import {Election, ElectionRaw} from "./model/election";
-import {AppListObservable} from "./app-list-observable";
-import {AppPromise} from "./app-promise";
-import {AppObjectObservable} from "./app-object-observable";
-import {AppListObservableObject} from "./app-list-observable-object";
-import {Party, PartyRaw} from "./model/party";
-import {District, DistrictRaw} from "./model/district";
+import {Election, ElectionRaw} from "../model/election";
+import {AppListObservable} from "../shared/app-list-observable";
+import {AppPromise} from "../shared/app-promise";
+import {AppObjectObservable} from "../shared/app-object-observable";
+import {AppListObservableObject} from "../shared/app-list-observable-object";
+import {Party, PartyRaw} from "../model/party";
+import {District} from "../model/district";
 import {Subscription} from "rxjs";
-import {RegionRaw} from "./model/region";
+import {RegionRaw} from "../model/region";
 import {AngularFire, AngularFireDatabase} from "angularfire2";
 import {DaoRegion} from "./dao-region";
 import {DaoDistrict} from "./dao-district";
