+++ src/app/dao/server-endpoint/dao-party.ts
@@ -1,9 +1,9 @@
-import {Election} from "./model/election";
-import {AppListObservable} from "./app-list-observable";
-import {AppPromise} from "./app-promise";
-import {AppObjectObservable} from "./app-object-observable";
-import {AppListObservableObject} from "./app-list-observable-object";
-import {Party, PartyRaw} from "./model/party";
+import {Election} from "../model/election";
+import {AppListObservable} from "../shared/app-list-observable";
+import {AppPromise} from "../shared/app-promise";
+import {AppObjectObservable} from "../shared/app-object-observable";
+import {AppListObservableObject} from "../shared/app-list-observable-object";
+import {Party, PartyRaw} from "../model/party";
 import {AngularFire, AngularFireDatabase} from "angularfire2";
 import {DaoElection} from "./dao-election";
 /**
