+++ src/app/dao/server-endpoint/dao-vote-count.ts
@@ -1,13 +1,13 @@
-import {Election, ElectionRaw} from "./model/election";
-import {AppListObservable} from "./app-list-observable";
-import {AppPromise} from "./app-promise";
-import {AppObjectObservable} from "./app-object-observable";
-import {AppListObservableObject} from "./app-list-observable-object";
-import {DistrictRaw} from "./model/district";
+import {Election, ElectionRaw} from "../model/election";
+import {AppListObservable} from "../shared/app-list-observable";
+import {AppPromise} from "../shared/app-promise";
+import {AppObjectObservable} from "../shared/app-object-observable";
+import {AppListObservableObject} from "../shared/app-list-observable-object";
+import {DistrictRaw} from "../model/district";
 import {Subscription} from "rxjs";
 import {AngularFire, AngularFireDatabase} from "angularfire2";
-import {VoteCount, VoteCountRaw} from "./model/vote-count";
-import {VoteType} from "./model/vote-type";
+import {VoteCount, VoteCountRaw} from "../model/vote-count";
+import {VoteType} from "../model/vote-type";
 import {DaoRegion} from "./dao-region";
 import {DaoDistrict} from "./dao-district";
 import {DaoElection} from "./dao-election";
